requirements.txt
alembic==1.13.1
aniso8601==9.0.1
appdirs==1.4.4
APScheduler==3.10.4
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.7.0
boto3==1.34.82
botocore==1.34.82
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cloudpickle==3.0.0
comm==0.1.3
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
dgl==0.9.1.post1
docker==7.0.0
entrypoints==0.4
executing==1.2.0
Flask==3.0.2
fonttools==4.44.1
future==0.18.3
gensim==3.8.3
gitdb==4.0.11
GitPython==3.1.42
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==21.2.0
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.6.0
importlib-resources==6.1.1
ipykernel==6.22.0
ipython==8.12.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.5
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.3
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mlflow==2.11.1
nest-asyncio==1.5.6
networkx @ file:///croot/networkx_1678964333703/work
nltk==3.8.1
node2vec==0.4.6
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
opencv-python==4.9.0.80
packaging @ file:///croot/packaging_1678965309396/work
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.5.0
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
prompt-toolkit==3.0.38
protobuf==4.25.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pure-sasl==0.6.2
pyarrow==12.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.15.1
PyHive==0.6.5
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.0.2
querystring-parser==1.2.4
regex==2023.5.5
requests @ file:///croot/requests_1682607517574/work
s3transfer==0.10.1
sasl==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
slack_sdk==3.27.1
smart-open==6.3.0
smmap==5.0.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
stack-data==0.6.2
threadpoolctl==3.1.0
thrift==0.16.0
thrift-sasl==0.4.3
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tornado==6.3.1
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets==5.9.0
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==5.2
urllib3 @ file:///croot/urllib3_1680254681959/work
wcwidth==0.2.6
Werkzeug==3.0.1
zipp==3.15.0
